# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GudangPro Inventory Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: getTransactions
      tags: [transactions]
      summary: List transactions with filters
      parameters:
        - name: start
          in: query
          schema:
            type: string
        - name: end
          in: query
          schema:
            type: string
        - name: warehouseId
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [IN, OUT, TRANSFER, ADJUSTMENT]
        - name: itemId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a transaction and update stock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a transaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    put:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update transaction (revert old stock, apply new)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      operationId: removeTransaction
      tags: [transactions]
      summary: Delete transaction and revert stock
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK

  # ─── REJECT ──────────────────────────────────────────────────────────────────
  /reject/outlets:
    get:
      operationId: getRejectOutlets
      tags: [reject]
      summary: List reject outlets
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RejectOutlet"
    post:
      operationId: createRejectOutlet
      tags: [reject]
      summary: Create reject outlet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectOutletInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RejectOutlet"
  /reject/master-items:
    get:
      operationId: getRejectMasterItems
      tags: [reject]
      summary: List reject master items
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RejectMasterItem"
    post:
      operationId: upsertRejectMasterItem
      tags: [reject]
      summary: Create or update reject master item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectMasterItemInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RejectMasterItem"
  /reject/master-items/{id}:
    delete:
      operationId: deleteRejectMasterItem
      tags: [reject]
      summary: Delete reject master item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /reject/batches:
    get:
      operationId: getRejectBatches
      tags: [reject]
      summary: List reject batches
      parameters:
        - name: start
          in: query
          schema:
            type: string
        - name: end
          in: query
          schema:
            type: string
        - name: outlet
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RejectBatch"
    post:
      operationId: createRejectBatch
      tags: [reject]
      summary: Create a reject batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectBatchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RejectBatch"
  /reject/batches/{id}:
    delete:
      operationId: deleteRejectBatch
      tags: [reject]
      summary: Delete reject batch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK

  # ─── MUSIC ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    UnitConversion:
      type: object
      required: [unitName, conversionRatio, operator]
      properties:
        unitName:
          type: string
        conversionRatio:
          type: number
        operator:
          type: string
          enum: ["*", "/"]

    # ─── ITEM ──────────────────────────────────────────────────────────────────
...
    TransactionItem:
      type: object
      required: [itemId, itemName, itemCode, qty, unit, conversionRatio, baseQty]
      properties:
        id:
          type: ["integer", "null"]
        itemId:
          type: string
        itemName:
          type: string
        itemCode:
          type: string
        qty:
          type: number
        unit:
          type: string
        conversionRatio:
          type: number
        baseQty:
          type: number
        note:
          type: ["string", "null"]
    TransactionItemInput:
      type: object
      required: [itemId, qty, unit, conversionRatio, baseQty]
      properties:
        itemId:
          type: string
        qty:
          type: number
        unit:
          type: string
        conversionRatio:
          type: number
        operator:
          type: string
          enum: ["*", "/"]
        baseQty:
          type: number
        note:
          type: ["string", "null"]
    Transaction:
      type: object
      required: [id, referenceNo, type, date, sourceWarehouseId, sourceWarehouseName, items]
      properties:
        id:
          type: string
        referenceNo:
          type: string
        type:
          type: string
          enum: [IN, OUT, TRANSFER, ADJUSTMENT]
        date:
          type: string
        sourceWarehouseId:
          type: string
        sourceWarehouseName:
          type: string
        targetWarehouseId:
          type: ["string", "null"]
        targetWarehouseName:
          type: ["string", "null"]
        partnerId:
          type: ["string", "null"]
        partnerName:
          type: ["string", "null"]
        deliveryOrderNo:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/TransactionItem"
    TransactionInput:
      type: object
      required: [type, date, sourceWarehouseId, items]
      properties:
        referenceNo:
          type: ["string", "null"]
        type:
          type: string
          enum: [IN, OUT, TRANSFER, ADJUSTMENT]
        date:
          type: string
        sourceWarehouseId:
          type: string
        targetWarehouseId:
          type: ["string", "null"]
        partnerId:
          type: ["string", "null"]
        partnerName:
          type: ["string", "null"]
        deliveryOrderNo:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/TransactionItemInput"
    TransactionUpdate:
      type: object
      required: [type, date, sourceWarehouseId, items]
      properties:
        referenceNo:
          type: ["string", "null"]
        type:
          type: string
          enum: [IN, OUT, TRANSFER, ADJUSTMENT]
        date:
          type: string
        sourceWarehouseId:
          type: string
        targetWarehouseId:
          type: ["string", "null"]
        partnerId:
          type: ["string", "null"]
        partnerName:
          type: ["string", "null"]
        deliveryOrderNo:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/TransactionItemInput"

    # ─── REJECT ────────────────────────────────────────────────────────────────
    RejectOutlet:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
    RejectOutletInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
    RejectMasterItem:
      type: object
      required: [id, code, name, baseUnit]
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        baseUnit:
          type: string
        conversions:
          type: array
          items:
            $ref: "#/components/schemas/UnitConversion"
    RejectMasterItemInput:
      type: object
      required: [code, name, baseUnit]
      properties:
        id:
          type: ["string", "null"]
        code:
          type: string
        name:
          type: string
        baseUnit:
          type: string
        conversions:
          type: array
          items:
            $ref: "#/components/schemas/UnitConversion"
    RejectBatchItem:
      type: object
      required: [qty, unit, baseQty]
      properties:
        id:
          type: ["integer", "null"]
        itemId:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        qty:
          type: number
        unit:
          type: string
        baseQty:
          type: number
        reason:
          type: ["string", "null"]
        inputQty:
          type: ["number", "null"]
        inputUnit:
          type: ["string", "null"]
    RejectBatchItemInput:
      type: object
      required: [qty, unit, baseQty]
      properties:
        itemId:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        qty:
          type: number
        unit:
          type: string
        baseQty:
          type: number
        reason:
          type: ["string", "null"]
        inputQty:
          type: ["number", "null"]
        inputUnit:
          type: ["string", "null"]
    RejectBatch:
      type: object
      required: [id, date, outlet, createdAt, items]
      properties:
        id:
          type: string
        date:
          type: string
        outlet:
          type: string
        createdAt:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/RejectBatchItem"
    RejectBatchInput:
      type: object
      required: [date, outlet, items]
      properties:
        date:
          type: string
        outlet:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/RejectBatchItemInput"

    # ─── MUSIC ─────────────────────────────────────────────────────────────────
...