# lib/api-zod/src/generated/types/rejectMasterItemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GudangPro Inventory Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UnitConversion } from './unitConversion';

export interface RejectMasterItemInput {
  /** @nullable */
  id?: string | null;
  code: string;
  name: string;
  baseUnit: string;
  conversions?: UnitConversion[];
}