# lib/api-zod/src/generated/types/transactionUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GudangPro Inventory Management API
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionItemInput } from './transactionItemInput';
import type { TransactionUpdateType } from './transactionUpdateType';

export interface TransactionUpdate {
  /** @nullable */
  referenceNo?: string | null;
  type: TransactionUpdateType;
  date: string;
  sourceWarehouseId: string;
  /** @nullable */
  targetWarehouseId?: string | null;
  /** @nullable */
  partnerId?: string | null;
  /** @nullable */
  partnerName?: string | null;
  /** @nullable */
  deliveryOrderNo?: string | null;
  /** @nullable */
  notes?: string | null;
  items: TransactionItemInput[];
}